+++ src/parser/giganto_import/tests.rs
@@ -14,11 +14,16 @@ use giganto_client::ingest::{
 
 use super::TryFromGigantoRecord;
 
+const RFC3339_ZERO: &str = "1970-01-01T00:00:00.000000000+00:00";
+
 #[test]
 fn giganto_conn() {
-    let data = "1669735962.571151000	localhost	fe80::2267:7cff:fef0:cb09	133	ff02::2	134	1	sf	0.000000000	0	-	0	0	1	0	21515	27889";
+    let data = format!(
+        "1669735962.571151000\tlocalhost\tfe80::2267:7cff:fef0:cb09\t133\t\
+         ff02::2\t134\t1\tsf\t{RFC3339_ZERO}\t0\t-\t0\t0\t1\t0\t21515\t27889"
+    );
 
-    let rec = stringrecord(data);
+    let rec = stringrecord(&data);
 
     assert!(Conn::try_from_giganto_record(&rec).is_ok());
 }
@@ -45,18 +50,8 @@ fn giganto_conn_rfc3339_start_time() {
 }
 
 #[test]
-fn giganto_conn_legacy_epoch_decimal_start_time() {
-    let data = "1669735962.571151000	localhost	fe80::2267:7cff:fef0:cb09	133\t\
-                ff02::2\t134\t1\tsf\t1669735962.571151000\t0\t-\t0\t0\t1\t0\t21515\t27889";
-    let rec = stringrecord(data);
-
-    let (conn, _) = Conn::try_from_giganto_record(&rec).unwrap();
-    assert_eq!(conn.start_time, 1_669_735_962_571_151_000);
-}
-
-#[test]
 fn giganto_http() {
-    let data = "1669773412.241856000	localhost	129.204.40.54	47697	218.144.35.150	80	0	0.000000000	0	1	0	21515	27889	GET	218.144.35.150	/root11.php	-	1.1	Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/68.0.3440.106 Safari/537.36	0	286	302	Found	-	-	-	-	-	-	-	-	10,10,10	-";
+    let data = "1669773412.241856000	localhost	129.204.40.54	47697	218.144.35.150	80	0	1970-01-01T00:00:00.000000000+00:00	0	1	0	21515	27889	GET	218.144.35.150	/root11.php	-	1.1	Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/68.0.3440.106 Safari/537.36	0	286	302	Found	-	-	-	-	-	-	-	-	10,10,10	-";
 
     let rec = stringrecord(data);
 
@@ -65,7 +60,7 @@ fn giganto_http() {
 
 #[test]
 fn giganto_rdp() {
-    let data = "1669775611.098308000	localhost	112.160.137.136	61572	103.153.182.151	3389	0	0.000000000	0	1	0	21515	27889	hello";
+    let data = "1669775611.098308000	localhost	112.160.137.136	61572	103.153.182.151	3389	0	1970-01-01T00:00:00.000000000+00:00	0	1	0	21515	27889	hello";
 
     let rec = stringrecord(data);
 
@@ -74,7 +69,7 @@ fn giganto_rdp() {
 
 #[test]
 fn giganto_smtp() {
-    let data = "1669136154.131718000	localhost	220.73.219.213	51280	67.195.204.72	25	0	0.000000000	0	1	0	21515	27889	hanjinyea@monami.com	-	-	-	-	-	-";
+    let data = "1669136154.131718000	localhost	220.73.219.213	51280	67.195.204.72	25	0	1970-01-01T00:00:00.000000000+00:00	0	1	0	21515	27889	hanjinyea@monami.com	-	-	-	-	-	-";
 
     let rec = stringrecord(data);
 
@@ -83,7 +78,7 @@ fn giganto_smtp() {
 
 #[test]
 fn giganto_dns() {
-    let data = "1664549996.073650000	collect	59.18.121.131	28116	211.252.150.11	53	17	0.000000000	0	1	0	21515	27889	discoplexa4.pl	-	30565	0	C_INTERNET	TXT	0	false	false	true	false	0";
+    let data = "1664549996.073650000	collect	59.18.121.131	28116	211.252.150.11	53	17	1970-01-01T00:00:00.000000000+00:00	0	1	0	21515	27889	discoplexa4.pl	-	30565	0	C_INTERNET	TXT	0	false	false	true	false	0";
 
     let rec = stringrecord(data);
 
@@ -92,7 +87,7 @@ fn giganto_dns() {
 
 #[test]
 fn giganto_malformed_dns() {
-    let data = "1761194588.611804000	localhost	127.0.0.1	46378	31.3.245.133	80	17	1761194588.611826000	1	1	2	100	200	1	42	1	1	0	0	1	1	16	32	[[65, 78, 61, 6d, 70, 6c, 65, 2e, 63, 6f, 6d]]	[[c0, c]]";
+    let data = "1761194588.611804000	localhost	127.0.0.1	46378	31.3.245.133	80	17	2025-10-23T04:43:08.611826000+00:00	1	1	2	100	200	1	42	1	1	0	0	1	1	16	32	[[65, 78, 61, 6d, 70, 6c, 65, 2e, 63, 6f, 6d]]	[[c0, c]]";
 
     let rec = stringrecord(data);
 
@@ -101,7 +96,7 @@ fn giganto_malformed_dns() {
 
 #[test]
 fn giganto_ntlm() {
-    let data = "1614130258.669753000	localhost	10.200.90.100	59271	192.168.0.111	445	0	0.000000000	0	1	0	21515	27889	-	it	-	-	-";
+    let data = "1614130258.669753000	localhost	10.200.90.100	59271	192.168.0.111	445	0	1970-01-01T00:00:00.000000000+00:00	0	1	0	21515	27889	-	it	-	-	-";
 
     let rec = stringrecord(data);
 
@@ -110,7 +105,7 @@ fn giganto_ntlm() {
 
 #[test]
 fn giganto_kerberos() {
-    let data = "1562093132.125665000	localhost	89.248.167.131	24067	210.117.142.55	88	0	0.000000000	0	1	0	21515	27889	0.000000000	0.000000000	1	client_realm	1	cname1,cname2	realm	1	sname1,sname2";
+    let data = "1562093132.125665000	localhost	89.248.167.131	24067	210.117.142.55	88	0	1970-01-01T00:00:00.000000000+00:00	0	1	0	21515	27889	1970-01-01T00:00:00.000000000+00:00	1970-01-01T00:00:00.000000000+00:00	1	client_realm	1	cname1,cname2	realm	1	sname1,sname2";
 
     let rec = stringrecord(data);
 
@@ -138,7 +133,7 @@ fn giganto_kerberos_rfc3339_datetimes() {
 
 #[test]
 fn giganto_ssh() {
-    let data = "1562093121.802019000	localhost	114.249.237.38	41260	203.254.132.18	22	0	0.000000000	0	1	0	21515	27889	SSH-2.0-Go	SSH-1.99-Cisco-1.25	aes128-cbc	hmac-sha1	none	diffie-hellman-group1-sha1	ssh-rsa	-	-	-	-	-	-";
+    let data = "1562093121.802019000	localhost	114.249.237.38	41260	203.254.132.18	22	0	1970-01-01T00:00:00.000000000+00:00	0	1	0	21515	27889	SSH-2.0-Go	SSH-1.99-Cisco-1.25	aes128-cbc	hmac-sha1	none	diffie-hellman-group1-sha1	ssh-rsa	-	-	-	-	-	-";
 
     let rec = stringrecord(data);
 
@@ -147,7 +142,7 @@ fn giganto_ssh() {
 
 #[test]
 fn giganto_dce_rpc() {
-    let data = "1614130373.991064000	localhost	192.168.0.111	58459	192.168.0.7	49670	6	1614130373.991064000	0	1	0	21515	27889	(0,0883AFE11F5DC91191A408002B14A0FA,3,0,045D888AEB1CC9119FE808002B104860,2,0,0,0),(1,1234567890ABCDEF1234567890ABCDEF,1,2,FEDCBA0987654321FEDCBA0987654321,3,4,5,6)	0:0,1:7";
+    let data = "1614130373.991064000	localhost	192.168.0.111	58459	192.168.0.7	49670	6	2021-02-24T01:32:53.991064000+00:00	0	1	0	21515	27889	(0,0883AFE11F5DC91191A408002B14A0FA,3,0,045D888AEB1CC9119FE808002B104860,2,0,0,0),(1,1234567890ABCDEF1234567890ABCDEF,1,2,FEDCBA0987654321FEDCBA0987654321,3,4,5,6)	0:0,1:7";
 
     let rec = stringrecord(data);
 
@@ -156,7 +151,7 @@ fn giganto_dce_rpc() {
 
 #[test]
 fn giganto_ftp() {
-    let data = r"1614130373.991064000	localhost	192.168.0.111	58459	192.168.0.7	49670	6	1614130373.991064000	0	1	0	21515	27889	anonymous	ftp@example.com	(EPSV,229,Entering Extended Passive Mode,true,192.168.4.76,196.216.2.24,31746,ftp://192.168.0.7/pub/stats/afrinic/delegated-afrinic-extended-latest.md5,74,226)";
+    let data = r"1614130373.991064000	localhost	192.168.0.111	58459	192.168.0.7	49670	6	2021-02-24T01:32:53.991064000+00:00	0	1	0	21515	27889	anonymous	ftp@example.com	(EPSV,229,Entering Extended Passive Mode,true,192.168.4.76,196.216.2.24,31746,ftp://192.168.0.7/pub/stats/afrinic/delegated-afrinic-extended-latest.md5,74,226)";
 
     let rec = stringrecord(data);
 
@@ -165,7 +160,7 @@ fn giganto_ftp() {
 
 #[test]
 fn giganto_ftp_reply_230() {
-    let data = r"1614130373.991064000	localhost	192.168.0.111	21	192.168.0.7	49670	6	1614130373.991064000	0	1	0	21515	27889	csanders	echo	(USER,230,User logged in, proceed,false,192.168.0.111,192.168.0.7,20,ftp://192.168.0.7/,0,1614130373991064000)";
+    let data = r"1614130373.991064000	localhost	192.168.0.111	21	192.168.0.7	49670	6	2021-02-24T01:32:53.991064000+00:00	0	1	0	21515	27889	csanders	echo	(USER,230,User logged in, proceed,false,192.168.0.111,192.168.0.7,20,ftp://192.168.0.7/,0,1614130373991064000)";
 
     let rec = stringrecord(data);
 
@@ -174,7 +169,7 @@ fn giganto_ftp_reply_230() {
 
 #[test]
 fn giganto_mqtt() {
-    let data = "1614130373.991064000	localhost	192.168.0.111	58459	192.168.0.7	49670	6	0.000000000	0	1	0	21515	27889	mqtt	3	my_client_id	10	topic1,topic2	10,10,10";
+    let data = "1614130373.991064000	localhost	192.168.0.111	58459	192.168.0.7	49670	6	1970-01-01T00:00:00.000000000+00:00	0	1	0	21515	27889	mqtt	3	my_client_id	10	topic1,topic2	10,10,10";
 
     let rec = stringrecord(data);
 
@@ -183,7 +178,7 @@ fn giganto_mqtt() {
 
 #[test]
 fn giganto_ldap() {
-    let data = "1614130373.991064000	localhost	192.168.0.111	58459	192.168.0.7	49670	0	0.000000000	0	1	0	21515	27889	2	3	opcode	result	diagnostic_mgs	object	argument";
+    let data = "1614130373.991064000	localhost	192.168.0.111	58459	192.168.0.7	49670	0	1970-01-01T00:00:00.000000000+00:00	0	1	0	21515	27889	2	3	opcode	result	diagnostic_mgs	object	argument";
 
     let rec = stringrecord(data);
 
@@ -192,7 +187,7 @@ fn giganto_ldap() {
 
 #[test]
 fn giganto_tls() {
-    let data = "1614130373.991064000	localhost	192.168.0.111	58459	192.168.0.7	49670	0	0.000000000	0	1	0	21515	27889	server_name	alpn_protocol	ja3	version	771,769,770	0,1,2	10	0,1	ja3s	serial	sub_country	sub_org_name	sub_comm_name	1700000000.000000000	1800000000.000000000	sub_alt_name	issuer_country	issuer_org_name	issuer_org_unit_name	issuer_common_name	10";
+    let data = "1614130373.991064000	localhost	192.168.0.111	58459	192.168.0.7	49670	0	1970-01-01T00:00:00.000000000+00:00	0	1	0	21515	27889	server_name	alpn_protocol	ja3	version	771,769,770	0,1,2	10	0,1	ja3s	serial	sub_country	sub_org_name	sub_comm_name	2023-11-14T22:13:20.000000000+00:00	2027-01-15T08:00:00.000000000+00:00	sub_alt_name	issuer_country	issuer_org_name	issuer_org_unit_name	issuer_common_name	10";
 
     let rec = stringrecord(data);
 
@@ -225,24 +220,8 @@ fn giganto_tls_rfc3339_datetimes() {
 }
 
 #[test]
-fn giganto_tls_legacy_epoch_decimal_validity() {
-    let data = "1614130373.991064000	localhost	192.168.0.111	58459	192.168.0.7\t\
-                49670\t0\t0.000000000\t0\t1\t0\t21515\t27889\tserver_name\t\
-                alpn_protocol\tja3\tversion\t771,769,770\t0,1,2\t10\t0,1\tja3s\t\
-                serial\tsub_country\tsub_org_name\tsub_comm_name\t\
-                1700000000.000000000\t1800000000.000000000\tsub_alt_name\t\
-                issuer_country\tissuer_org_name\tissuer_org_unit_name\t\
-                issuer_common_name\t10";
-    let rec = stringrecord(data);
-
-    let (tls, _) = Tls::try_from_giganto_record(&rec).unwrap();
-    assert_eq!(tls.validity_not_before, 1_700_000_000_000_000_000);
-    assert_eq!(tls.validity_not_after, 1_800_000_000_000_000_000);
-}
-
-#[test]
 fn giganto_smb() {
-    let data = "1614130373.991064000	localhost	192.168.0.111	58459	192.168.0.7	49670	0	0.000000000	0	1	0	21515	27889	0	path	service	file_name	10	20	30	10000000	20000000	10000000	20000000";
+    let data = "1614130373.991064000	localhost	192.168.0.111	58459	192.168.0.7	49670	0	1970-01-01T00:00:00.000000000+00:00	0	1	0	21515	27889	0	path	service	file_name	10	20	30	10000000	20000000	10000000	20000000";
 
     let rec = stringrecord(data);
 
@@ -251,7 +230,7 @@ fn giganto_smb() {
 
 #[test]
 fn giganto_nfs() {
-    let data = "1614130373.991064000	localhost	192.168.0.111	58459	192.168.0.7	49670	0	0.000000000	0	1	0	21515	27889	-	-";
+    let data = "1614130373.991064000	localhost	192.168.0.111	58459	192.168.0.7	49670	0	1970-01-01T00:00:00.000000000+00:00	0	1	0	21515	27889	-	-";
 
     let rec = stringrecord(data);
 
@@ -260,7 +239,7 @@ fn giganto_nfs() {
 
 #[test]
 fn giganto_bootp() {
-    let data = "1614130373.991064000	localhost	192.168.0.111	58459	192.168.0.7	49670	6	0.000000000	0	1	0	21515	27889	0	1	2	3	192.168.4.1	192.168.4.2	192.168.4.3	192.168.4.4	0,1,2	sname	file";
+    let data = "1614130373.991064000	localhost	192.168.0.111	58459	192.168.0.7	49670	6	1970-01-01T00:00:00.000000000+00:00	0	1	0	21515	27889	0	1	2	3	192.168.4.1	192.168.4.2	192.168.4.3	192.168.4.4	0,1,2	sname	file";
 
     let rec = stringrecord(data);
 
@@ -269,7 +248,7 @@ fn giganto_bootp() {
 
 #[test]
 fn giganto_dhcp() {
-    let data = "1614130373.991064000	localhost	192.168.0.111	58459	192.168.0.7	49670	6	0.000000000	0	1	0	21515	27889	0	192.168.4.1	192.168.4.2	192.168.4.3	192.168.4.4	192.168.4.5	192.168.4.11,192.168.4.22	192.168.4.33,192.168.4.44	192.168.4.6	1	192.168.4.7	0,1,2	message	1	1	0,1,2	1	0,1,2	53:05,51:00015180";
+    let data = "1614130373.991064000	localhost	192.168.0.111	58459	192.168.0.7	49670	6	1970-01-01T00:00:00.000000000+00:00	0	1	0	21515	27889	0	192.168.4.1	192.168.4.2	192.168.4.3	192.168.4.4	192.168.4.5	192.168.4.11,192.168.4.22	192.168.4.33,192.168.4.44	192.168.4.6	1	192.168.4.7	0,1,2	message	1	1	0,1,2	1	0,1,2	53:05,51:00015180";
 
     let rec = stringrecord(data);
 
@@ -280,7 +259,7 @@ fn giganto_dhcp() {
     );
 
     // Empty options
-    let data_empty = "1614130373.991064000	localhost	192.168.0.111	58459	192.168.0.7	49670	6	0.000000000	0	1	0	21515	27889	0	192.168.4.1	192.168.4.2	192.168.4.3	192.168.4.4	192.168.4.5	192.168.4.11,192.168.4.22	192.168.4.33,192.168.4.44	192.168.4.6	1	192.168.4.7	0,1,2	message	1	1	0,1,2	1	0,1,2	-";
+    let data_empty = "1614130373.991064000	localhost	192.168.0.111	58459	192.168.0.7	49670	6	1970-01-01T00:00:00.000000000+00:00	0	1	0	21515	27889	0	192.168.4.1	192.168.4.2	192.168.4.3	192.168.4.4	192.168.4.5	192.168.4.11,192.168.4.22	192.168.4.33,192.168.4.44	192.168.4.6	1	192.168.4.7	0,1,2	message	1	1	0,1,2	1	0,1,2	-";
     let rec_empty = stringrecord(data_empty);
     let (dhcp_empty, _) = Dhcp::try_from_giganto_record(&rec_empty).unwrap();
     assert!(dhcp_empty.options.is_empty());
@@ -288,7 +267,7 @@ fn giganto_dhcp() {
 
 #[test]
 fn giganto_dhcp_invalid_options() {
-    let base = "1614130373.991064000\tlocalhost\t192.168.0.111\t58459\t192.168.0.7\t49670\t6\t0.000000000\t0\t1\t0\t21515\t27889\t0\t192.168.4.1\t192.168.4.2\t192.168.4.3\t192.168.4.4\t192.168.4.5\t192.168.4.11,192.168.4.22\t192.168.4.33,192.168.4.44\t192.168.4.6\t1\t192.168.4.7\t0,1,2\tmessage\t1\t1\t0,1,2\t1\t0,1,2\t";
+    let base = "1614130373.991064000\tlocalhost\t192.168.0.111\t58459\t192.168.0.7\t49670\t6\t1970-01-01T00:00:00.000000000+00:00\t0\t1\t0\t21515\t27889\t0\t192.168.4.1\t192.168.4.2\t192.168.4.3\t192.168.4.4\t192.168.4.5\t192.168.4.11,192.168.4.22\t192.168.4.33,192.168.4.44\t192.168.4.6\t1\t192.168.4.7\t0,1,2\tmessage\t1\t1\t0,1,2\t1\t0,1,2\t";
 
     // Missing colon separator
     let rec = stringrecord(&format!("{base}invalid"));
@@ -309,7 +288,7 @@ fn giganto_dhcp_invalid_options() {
 
 #[test]
 fn giganto_radius() {
-    let data = "1756197618.963374000	localhost	127.0.0.1	53031	192.0.2.1	1812	17	1440447766.441298000	0	1	0	21515	27889	103	1	255	40b664dbf5d681b2adbd1769515118c8		115,116,101,118,101	219,198,196,183,88,190,20,240,5,179,135,124,158,47,182,1	-	192.168.0.28	123	-	-	0	";
+    let data = "1756197618.963374000	localhost	127.0.0.1	53031	192.0.2.1	1812	17	2015-08-24T20:22:46.441298000+00:00	0	1	0	21515	27889	103	1	255	40b664dbf5d681b2adbd1769515118c8		115,116,101,118,101	219,198,196,183,88,190,20,240,5,179,135,124,158,47,182,1	-	192.168.0.28	123	-	-	0	";
 
     let rec = stringrecord(data);
 
@@ -318,7 +297,7 @@ fn giganto_radius() {
 
 #[test]
 fn giganto_icmp() {
-    let data = "1756197618.963374000	localhost	192.168.1.1	192.168.1.2	1	1000000000.000000000	0	1	1	100	100	8	0	1234	1	56	[8, 0, ff, ff]";
+    let data = "1756197618.963374000	localhost	192.168.1.1	192.168.1.2	1	2001-09-09T01:46:40.000000000+00:00	0	1	1	100	100	8	0	1234	1	56	[8, 0, ff, ff]";
 
     let rec = stringrecord(data);
 
@@ -353,7 +332,7 @@ fn sysmon_process_create_sample() {
 
 #[test]
 fn sysmon_file_create_time_sample() {
-    let data = "1691452807.978000000	sensor1	agent-a	agent-id-1	{11111111-2222-3333-4444-555555555555}	1234	C:\\Windows\\System32\\cmd.exe	C:\\Temp\\file.txt	1691452700.000000000	1691452600.000000000	DOMAIN\\User";
+    let data = "1691452807.978000000	sensor1	agent-a	agent-id-1	{11111111-2222-3333-4444-555555555555}	1234	C:\\Windows\\System32\\cmd.exe	C:\\Temp\\file.txt	2023-08-07T23:58:20.000000000+00:00	2023-08-07T23:56:40.000000000+00:00	DOMAIN\\User";
     let rec = stringrecord(data);
     assert!(FileCreationTimeChanged::try_from_giganto_record(&rec).is_ok());
 }
@@ -394,14 +373,14 @@ fn sysmon_process_terminate_sample() {
 
 #[test]
 fn sysmon_file_create_sample() {
-    let data = "1691452807.978000000	sensor1	agent-a	agent-id-1	{aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee}	4321	C:\\Windows\\System32\\svchost.exe	C:\\Temp\\created.txt	1691452700.000000000	DOMAIN\\User";
+    let data = "1691452807.978000000	sensor1	agent-a	agent-id-1	{aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee}	4321	C:\\Windows\\System32\\svchost.exe	C:\\Temp\\created.txt	2023-08-07T23:58:20.000000000+00:00	DOMAIN\\User";
     let rec = stringrecord(data);
     assert!(FileCreate::try_from_giganto_record(&rec).is_ok());
 }
 
 #[test]
 fn sysmon_file_create_stream_hash_sample() {
-    let data = "1691452807.978000000	sensor1	agent-a	agent-id-1	{aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee}	4321	C:\\Windows\\System32\\svchost.exe	C:\\Temp\\created.txt	1691452700.000000000	SHA256=ABCDEF,MD5=123456	contents	DOMAIN\\User";
+    let data = "1691452807.978000000	sensor1	agent-a	agent-id-1	{aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee}	4321	C:\\Windows\\System32\\svchost.exe	C:\\Temp\\created.txt	2023-08-07T23:58:20.000000000+00:00	SHA256=ABCDEF,MD5=123456	contents	DOMAIN\\User";
     let rec = stringrecord(data);
     assert!(FileCreateStreamHash::try_from_giganto_record(&rec).is_ok());
 }